Moved utility functions to own file.
